sync

+++ api/src/main/java/org/openmrs/api/impl/UserServiceImpl.java
@@ -1,15 +1,11 @@
 /**
- * The contents of this file are subject to the OpenMRS Public License
- * Version 1.0 (the "License"); you may not use this file except in
- * compliance with the License. You may obtain a copy of the License at
- * http://license.openmrs.org
+ * This Source Code Form is subject to the terms of the Mozilla Public License,
+ * v. 2.0. If a copy of the MPL was not distributed with this file, You can
+ * obtain one at http://mozilla.org/MPL/2.0/. OpenMRS is also distributed under
+ * the terms of the Healthcare Disclaimer located at http://openmrs.org/license.
  *
- * Software distributed under the License is distributed on an "AS IS"
- * basis, WITHOUT WARRANTY OF ANY KIND, either express or implied. See the
- * License for the specific language governing rights and limitations
- * under the License.
- *
- * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
+ * Copyright (C) OpenMRS Inc. OpenMRS is a registered trademark and the OpenMRS
+ * graphic logo is a trademark of OpenMRS Inc.
  */
 package org.openmrs.api.impl;
 
@@ -29,14 +25,16 @@ import org.openmrs.Privilege;
 import org.openmrs.PrivilegeListener;
 import org.openmrs.Role;
 import org.openmrs.User;
+import org.openmrs.annotation.Authorized;
+import org.openmrs.annotation.Logging;
 import org.openmrs.api.APIAuthenticationException;
 import org.openmrs.api.APIException;
+import org.openmrs.api.CannotDeleteRoleWithChildrenException;
 import org.openmrs.api.UserService;
 import org.openmrs.api.context.Context;
 import org.openmrs.api.db.DAOException;
 import org.openmrs.api.db.UserDAO;
 import org.openmrs.patient.impl.LuhnIdentifierValidator;
-import org.openmrs.util.OpenmrsConstants;
 import org.openmrs.util.OpenmrsUtil;
 import org.openmrs.util.PrivilegeConstants;
 import org.openmrs.util.RoleConstants;
@@ -46,7 +44,7 @@ import org.springframework.transaction.annotation.Transactional;
 /**
  * Default implementation of the user service. This class should not be used on its own. The current
  * OpenMRS implementation should be fetched from the Context
- *
+ * 
  * @see org.openmrs.api.UserService
  * @see org.openmrs.api.context.Context
  */
@@ -71,12 +69,14 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 	 * @see org.openmrs.api.UserService#createUser(org.openmrs.User, java.lang.String)
 	 * @deprecated
 	 */
+	@Deprecated
 	public User createUser(User user, String password) throws APIException {
 		return Context.getUserService().saveUser(user, password);
 	}
 	
 	/**
 	 * @see org.openmrs.api.UserService#saveUser(org.openmrs.User, java.lang.String)
+	 * @deprecated replaced by {@link #createUser(User, String)}
 	 */
 	public User saveUser(User user, String password) throws APIException {
 		if (user.getUserId() == null) {
@@ -90,7 +90,7 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 		// if we're creating a user and a password wasn't supplied, throw an
 		// error
 		if (user.getUserId() == null && (password == null || password.length() < 1)) {
-			throw new APIException("A password is required when creating a user");
+			throw new APIException("User.creating.password.required", (Object[]) null);
 		}
 		
 		if (hasDuplicateUsername(user)) {
@@ -259,7 +259,7 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 	 */
 	public void purgePrivilege(Privilege privilege) throws APIException {
 		if (OpenmrsUtil.getCorePrivileges().keySet().contains(privilege.getPrivilege())) {
-			throw new APIException("Cannot delete a core privilege");
+			throw new APIException("Privilege.cannot.delete.core", (Object[]) null);
 		}
 		
 		dao.deletePrivilege(privilege);
@@ -319,7 +319,11 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 		}
 		
 		if (OpenmrsUtil.getCoreRoles().keySet().contains(role.getRole())) {
-			throw new APIException("Cannot delete a core role");
+			throw new APIException("Role.cannot.delete.core", (Object[]) null);
+		}
+		
+		if (role.hasChildRoles()) {
+			throw new CannotDeleteRoleWithChildrenException();
 		}
 		
 		dao.deleteRole(role);
@@ -343,7 +347,9 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 	
 	/**
 	 * @see org.openmrs.api.UserService#changePassword(org.openmrs.User, java.lang.String)
+	 * @deprecated replaced by {@link #changePassword(User, String, String)}
 	 */
+	@Deprecated
 	public void changePassword(User u, String pw) throws APIException {
 		dao.changePassword(u, pw);
 	}
@@ -445,7 +451,7 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 	
 	/**
 	 * Convenience method to check if the authenticated user has all privileges they are giving out
-	 *
+	 * 
 	 * @param new user that has privileges
 	 */
 	private void checkPrivileges(User user) {
@@ -455,7 +461,7 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 		for (Role r : roles) {
 			if (r.getRole().equals(RoleConstants.SUPERUSER)
 			        && !Context.hasPrivilege(PrivilegeConstants.ASSIGN_SYSTEM_DEVELOPER_ROLE)) {
-				throw new APIException("You must have the role '" + RoleConstants.SUPERUSER + "' in order to assign it.");
+				throw new APIException("User.you.must.have.role", new Object[] { RoleConstants.SUPERUSER });
 			}
 			if (r.getPrivileges() != null) {
 				for (Privilege p : r.getPrivileges()) {
@@ -467,7 +473,7 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 		}
 		
 		if (requiredPrivs.size() == 1) {
-			throw new APIException("You must have privilege '" + requiredPrivs.get(0) + "' in order to assign it.");
+			throw new APIException("User.you.must.have.privilege", new Object[] { requiredPrivs.get(0) });
 		} else if (requiredPrivs.size() > 1) {
 			StringBuilder txt = new StringBuilder("You must have the following privileges in order to assign them: ");
 			for (String s : requiredPrivs) {
@@ -483,7 +489,7 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 	public User setUserProperty(User user, String key, String value) {
 		if (user != null) {
 			if (!Context.hasPrivilege(PrivilegeConstants.EDIT_USERS) && !user.equals(Context.getAuthenticatedUser())) {
-				throw new APIException("You are not authorized to change " + user.getUserId() + "'s properties");
+				throw new APIException("you.are.not.authorized.change.properties", new Object[] { user.getUserId() });
 			}
 			
 			user.setUserProperty(key, value);
@@ -509,7 +515,7 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 			// the user being edited is not the current user, throw an
 			// exception
 			if (!Context.hasPrivilege(PrivilegeConstants.EDIT_USERS) && !user.equals(Context.getAuthenticatedUser())) {
-				throw new APIException("You are not authorized to change " + user.getUserId() + "'s properties");
+				throw new APIException("you.are.not.authorized.change.properties", new Object[] { user.getUserId() });
 			}
 			
 			user.removeUserProperty(key);
@@ -528,7 +534,7 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 	
 	/**
 	 * Generates system ids based on the following algorithm scheme: user_id-check digit
-	 *
+	 * 
 	 * @see org.openmrs.api.UserService#generateSystemId()
 	 */
 	@Transactional(readOnly = true)
@@ -570,8 +576,8 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 	 * @see org.openmrs.api.UserService#purgeUser(org.openmrs.User, boolean)
 	 */
 	public void purgeUser(User user, boolean cascade) throws APIException {
-		if (cascade == true) {
-			throw new APIException("I don't think we want to cascade here");
+		if (cascade) {
+			throw new APIException("cascade.do.not.think", (Object[]) null);
 		}
 		
 		dao.deleteUser(user);
@@ -580,7 +586,7 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 	/**
 	 * Convenience method to check if the authenticated user has all privileges they are giving out
 	 * to the new role
-	 *
+	 * 
 	 * @param new user that has privileges
 	 */
 	private void checkPrivileges(Role role) {
@@ -693,7 +699,7 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 	public User saveUserProperty(String key, String value) {
 		User user = Context.getAuthenticatedUser();
 		if (user == null) {
-			throw new APIException("No Authenticated user found");
+			throw new APIException("no.authenticated.user.found", (Object[]) null);
 		}
 		user.setUserProperty(key, value);
 		return dao.saveUser(user, null);
@@ -703,13 +709,34 @@ public class UserServiceImpl extends BaseOpenmrsService implements UserService {
 	public User saveUserProperties(Map<String, String> properties) {
 		User user = Context.getAuthenticatedUser();
 		if (user == null) {
-			throw new APIException("No Authenticated user found");
+			throw new APIException("no.authenticated.user.found", (Object[]) null);
 		}
 		user.getUserProperties().clear();
 		for (Map.Entry<String, String> entry : properties.entrySet()) {
 			user.setUserProperty(entry.getKey(), entry.getValue());
 		}
 		return dao.saveUser(user, null);
+	}
+	
+	/**
+	 * @see UserService#changePassword(User, String, String)
+	 */
+	@Override
+	@Authorized(PrivilegeConstants.EDIT_USER_PASSWORDS)
+	@Logging(ignoredArgumentIndexes = { 1, 2 })
+	public void changePassword(User user, String oldPassword, String newPassword) throws APIException {
+		if (user.getUserId() == null) {
+			throw new APIException("user.must.exist", (Object[]) null);
+		}
+		if (oldPassword == null) {
+			if (!Context.hasPrivilege(PrivilegeConstants.EDIT_USER_PASSWORDS)) {
+				throw new APIException("null.old.password.privilege.required", (Object[]) null);
+			}
+		} else if (!dao.getLoginCredential(user).checkPassword(oldPassword)) {
+			throw new APIException("old.password.not.correct", (Object[]) null);
+		}
+		OpenmrsUtil.validatePassword(user.getUsername(), newPassword, user.getSystemId());
+		dao.changePassword(user, newPassword);
 	}
 	
 }
